1.锁仓创建单测 2.统计地址单测 3.举报数据补充，部分

+++ browser-agent/src/main/java/com/platon/browser/complement/service/param/converter/RestrictingCreateConverter.java
@@ -1,6 +1,7 @@
 package com.platon.browser.complement.service.param.converter;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -29,7 +30,7 @@ public class RestrictingCreateConverter extends BusinessParamConverter<Restricti
 				.address(account)
 				.amount(new BigDecimal(plan.getAmount()))
 				.epoch(plan.getEpoch().longValue())
-				.number(tx.getNum())
+				.number(BigInteger.valueOf(tx.getNum()))
 				.build();
 			}).collect(Collectors.toList());
